ARM: dts: msm: Enable fw_devlink in strict mode for Waipio
The FW devlink feature is used so that the kernel can establish
supplier-consumer relations between devices based on common
properties in the devicetree (e.g. IOMMUs, clocks, and power domains).

For example, if a peripheral in the devicetree uses a particular clock,
it can express this through the "clocks" property. The kernel will then
ensure that the clock device has been successfully bound to its driver
before the peripheral can attempt to bind to its own driver.

The kernel does not treat all of these dependencies as mandatory
however, such as a dependency on an IOMMU, and can allow IOMMU clients
to probe before their IOMMU has finished binding to its driver.

This is problematic for our systems, as clients upstream of an IOMMU
must have their stream IDs (SIDs) programmed into the IOMMU's stream
matching registers (SMRs), so that the IOMMU doesn't fault a client's
transactionas a transaction with an unidentified stream. Thus enable FW
devlink feature in strict mode for Waipio to ensure that IOMMU clients
can only probe after their IOMMU device has been bound to a driver.

Change-Id: I32d4bddb80706aa048cff6a73e68d82384313150