Revert "ARM: dts: msm: Remove VSMMU node for Kalama's VM"
This reverts commit f62a3430c4e51b490a4e6de4feefc34b8582f8c0. The
aforementioned commit was reverted as it was causing Kalama's trusted VM
to not boot.

This was root-caused to be a problem with the number of stream
matching registers (SMRs) and context banks (CBs) that was requested
from the trusted VM. The values that were specified exceeded the maximum
allowed number of SMRs and CBs for Kalama's VM, so the hypervisor
rejected the request for said resources, and failed to boot the VM.

The current maximum numbers of SMRs and CBs that are allowed for Kalama's
VM, are 4 and 2 respectively, so use those values. They can be increased
later on when the hypervisor allows it.

Change-Id: Iaa8711c1ada4398daea312e689a5c01be332357e