ARM: dts: msm: Increase GDSC timeouts for Kalama
It's been observed that sometimes the USB GDSCs take longer than 500us
to turn off. It's not clear how much longer they're taking, nor is it
clear if other GDSCs will eventually run into the same problem. So
increase the timeouts for all GDSCs until the requirements are further
characterized.

Change-Id: If836e141ad7c7f55ad295cdbbfb0d75741345e96