ARM: dts: msm: Add strong pull up config for applicable PINs
This change adds qcom,i2c_pull property for the applicable I2C GPIOs.
This is on top of the change 4020796 and as per the IPS shared data.
With this we need to remove bias-disable property during sleep state,
otherwise the strong PULL will not be applicable.

Note that for the boards where it has external pull up but also supports
internal pull up, use case owner/customer need to make sure that internal
pull up is not applied for active and bias-disable is applied for sleep
state. wherever external pull is not present and internal pull up is
supported, we gave commented DTSI config to ease enabling quickly.

Change-Id: Ie71e8466e28f1cce6b1c5d5132088bd3aeed530b