ARM: dts: msm: Add apss_ubwcp_pwr_ctrl GDSC for Pineapple
The apss_ubwcp_pwr_ctrl register behaves like a basic GDSC and when
enabled will vote for both clock and power required for UBWCP operation
and configuration. TZ and Hypervisor also have their own voting
registers for SMMU configuration, so don't poll for the GDSC to turn off
when disabling.

Change-Id: I4dbcd19df3bb2544cb425af05911e0602a825157