ARM: dts: msm: Vote MMNOC BW for cam_cc_titan_top_gdsc on Kalama
The cam_cc_titan_top_gdsc causes camera to permanently assert qactive
when it's enabled. If MMNOC attempts to collapse when the GDSC is
enabled, then the qactive signal will cause BCM to get stuck in the
collapse sequence. This will eventually lead to rpmh driver timeouts
when requested for MMNOC. Fix this by explicitly voting for MMNOC on
behalf of the GDSC.

And adding these interconnect phandles creates a dependency cycle
between the GDSC and camcc, so remove the GDSC phandle from camcc. The
camcc device only needed it to prevent warnings when enabling critical
clocks with the enable() callback during probe, but we can accomplish
the same thing by enabling the clocks manually in the probe function
without checking status.

Change-Id: I7772532bdf2d7443de02d6f40b48042920bf48d2