ARM: dts: qcom: add SPMI bus controller for Pineapple
Add spmi-pmic-arb devices for the primary and secondary SPMI buses
found on Pineapple.  The primary bus operates at 19.2 MHz and is
used for most of the PMICs.

The secondary bus operates at 4.8 MHz and is used exclusively for
charging PMICs.  Note that the secondary bus is not connected to
the SoC on the board due to voltage level differences.  Therefore,
keep the secondary bus device disabled.

Change-Id: Ida4a90a22278dd15729360f4aa54f44820ed6ac5