ARM: dts: msm: Add proxy-consumer for GDSCs on Pineapple
Add proxy-consumers for GDSCs that are enabled from bootloaders and are
critical for early kernel operation.

Additionally, add the USB3_GDSC-supply property to the usb0 device.
Without this, the USB driver doesn't vote for it and the GDSC will have
no consumers, which means it'll immediately turn off in sync-state.

Change-Id: If68566e52373f75ad4b306182c748cc8067f48be