ARM: dts: msm: Pull-up Wailua GPIO 7 to detect Vbus on cinder
Vdet pin of NX5P1100 switch is goes high Vbus is supplied to it.
This DET pin is open drain PMOS output with default state low.
And its connected to GPIO 7 of Wailua which is used for detection
of vbus. Here we need to pull up GPIO 7, so that DET can go high
when Vbus is detected on NX5P switch.

Change-Id: I81298ffb779f90ba169c9ce9758b4cde9d1c0276