dt-bindings: Add qcom,ignore-numpagendxb to ignore NUMPAGENDXB
We sometimes might want to ignore the SMMU global address space size
being reported to us in SMMU_IDR1.NUMPAGENDXB, and assume instead that
the TCU address space size we report in our device tree is equal twice
the size of the global address space (such that the context bank
address space and global address space are equal in size, with the
context bank address space taking the upper half). Using this
assumption, we can then use the DT-provided size (in the "reg ="
property) to determine the layout of the global and context bank
address spaces.

Define "qcom,ignore-numpagendxb" to allow users to allow this override
to occur. Move the parse_driver_options() call earlier into the probe
process since we need to read the options earlier now.

Change-Id: I3684715d1f80af425b6ce74264d40e676a29eef8