ARM: dts: msm: Use regulator-fixed to drive GPIO 43 on cinder
Use fixed voltage regulator framework to assert/deassert Vbus-out
pin (TLMM GPIO 43). And this tlmm pin is connected to EN pin of
NX5P1100 switch. This regualator phandle is added to usb-conn-gpio
node, so that enabling/disabling this regulator would assert or
deassert the GPIO. And thus Vbus can be supplied to the peripheral
using this GPIO.

Change-Id: Ic2c4c8a88d9c1808588bce8fd25a002ceb1dba20