ARM: dts: msm: Add USB support for Lemans
Add USB controller and PHY configuration for LEMANS RUMI.
Add primary, secondary and tertiary USB nodes on LEMANS.

Since primary controller is by default in peripheral, make A4 as
host mode only controller so that A6 is picked for UDC.

The SS phy parsing logic was refactored and now we
skip the last reads. We skip the delays as well.

Change-Id: I2931b0b57f64b420dffb22d65645a527ecbee351