ARM: dts: msm: Add wcd_usbss reference to USB on Pineapple QRD
On QRD platform the USB D+/D- signal lines are first routed through
the WCD939x USB subsystem before connecting to the USB controller.
Add a phandle to the former to the USB device node.  This will allow
the USB driver to control the D+/D- switches when USB 2.0 functionality
is needed.

Change-Id: I34640919e2dfb3ac8d4489491a64721829b21ba1