ARM: dts: msm: khaje: Add workaround for erratum 858921
ARM64_ERRATUM_858921 is presently broken if an affected
CPU is a secondary CPU. You may find more details here [1].

Till upstream converges to a acceptable solution, this
patch introduces device tree level workaround
"qcom,erratum-858921" for impacted QCOM SoCs.

The downside of course is that workaround gets applied to
non impacted CPUs as well on such systems. This would be
reverted with proper fix, once it is concluded at [1].

1. https://lore.kernel.org/linux-arm-kernel/20230113111648.1977473-1-maz@kernel.org/.

Change-Id: I17d0525870a17ec61dcff36a204d5e661a02cf8c