ARM: dts: qcom: add PMIC Glink ADC device for pineapple boards
Add a PMIC Glink ADC device along with ADC channel nodes for
pineapple MTP and QRD boards.  This exposes an interface for
software to read PMIC charger ADC channels for input current,
charge current, and die temperature.  This is useful for PMICs
connected via I2C which are only accessible to charger firmware.

Change-Id: Ifb43e35f95618b32df7f8e0350c8c6ed65e6efc3