ARM: dts: msm: Add oemvm qrtr gunyah node for pineapple
Add the nodes for enable qrtr communication between primary vm and
oemvm on pineapple.

This adds the platform devices and vdevice descriptions to start the
qrtr gunyah transport on both primary vm and oemvm device trees. This
also adds the device tree node to configure qrtr as node id 21 on oem
vm.

Change-Id: I302f74cf1887e4a9b733deac2c803a6b63e51311