ARM: dts: qcom: Add support for builds which don't support DT overlay
The current implementation of the device tree platform map requires
users to access the raw map and know internal details of its
structure. It also prevents the use of dynamic logic to modify the
map at build time for build requirements.

Expose the map via macros to support dynamic configuration of DT
overlay.

Change-Id: I3132e601b559abe3db140284259374c4c048ceed