ARM: dts: msm: Mark titan-top GDSC as parent-supply for cam cc GDSCs
titan-top GDSC needs to be on when its child GDSCs are on. Usually this
sequencing is taken care by consumers. If titan top is disabled before
its child GDSCs there is a chance of MMNOC hang due to MM0, MM1 BCM is
stuck as it has qactive signal from the GDSCs.
To fix above, for now mark titan-top GDSC as parent-supply to all child
cam cc GDSCs for pineapple.

Change-Id: I76e907a3c6e412c560c7128d78819e360b36268e