ARM: dts: msm: Add new core-hang info property
Current properties "qcom,threshold-arr" and "qcom,config-arr"
assumes that logical CPU number & physical CPU has 1:1
mapping and accordingly all the registers placed linearly with
increasing order of logical CPU numbers.
But if any CPU is not available then 1:1 mapping is broken
as we are having logically contiguous CPU numbers even if
physical cpus are not available.
To resolve this we are replacing old properties with a new property
"qcom,chd-percpu-info" and map core-hang registers wrt to CPU phandles.
The format is <&CPUx_Phandle CPUx_Threshold CPUx_Config>.

Change-Id: Id1d50a0c03ffb15979e2c3e7bb00cc8e98625b49