bindings: add support for qcom,qti-pmic-glink
The PMIC Glink device provides the interface for clients to
communicate over GLink for sending/receiving data to/from
charger firmware that runs on a remote subsystem (e.g. ADSP)
which supports charging and gauging. Add the bindings for it.

This is a snapshot taken as of qcom-6.1
commit 2927dc2d179b ("bindings: add support for
qcom,pmic-glink") which was then converted to yaml.

Rename qcom,pmic-glink.yaml to qcom,qti-pmic-glink.yaml to
avoid a conflict with the independent upstream implementation
of this file.

Change-Id: Ia4f5ff346ebfa773bc0fe3d0410626ea67e1da31
Signed-off-by: David Collins <quic_collinsd@quicinc.com>