bindings: Add altmode-glink
The altmode mode device provides an interface for Type-C alternate mode
clients to receive data such as Pin Assignment Notifications from the
Type-C stack running on a remote subsystem (e.g. DSP) via the PMIC GLINK
interface.

This is a snapshot taken as of qcom-6.1
commit 956e9008b1cf ("bindings: Add altmode-glink")
which was then converted to yaml.

Change-Id: Iec1cf5e64ed9b4eb93fbfa6aacc998d4f4f730e7
Signed-off-by: David Collins <quic_collinsd@quicinc.com>