dt-bindings: spmi: add SPMI PMIC arbiter debug bus bindings
Add binding documentation for the SPMI debug bus found on SPMI
PMIC arbiter version 5 and above.  This debug bus has read and
write access to all PMIC peripherals regardless of ownership
configurations.  It cannot be used on production devices because
it is disabled by an eFuse.

This is a snapshot of the file qcom,spmi-pmic-arb-debug.txt
taken as of qcom-6.1 commit b522c3b6d065 ("dt-bindings: spmi:
spmi-pmic-arb-debug: define enable fuse property")
which was then converted to yaml.

Change-Id: I344ddea23ecb09bede5cbdb19197ab52b24dda44
Signed-off-by: David Collins <quic_collinsd@quicinc.com>