ARM: dts: msm: Add support for GCC, GPUCC, TCSRCC on cliffs
Add support for GCC, GPUCC and TCSRCC clock controller nodes on
CLIFFS platform. While at it, move the GDSC's of GCC and GPUCC to
real from dummy.

Also remove support for bi_tcxo and bi_tcxo_ao nodes as they are not
required since we can directly use rpmhcc phandle which is a fixed-clock
to get 19.2MHz frequency.

Change-Id: Ifa408eb327e2cc643095a5a6985e34bbf3d85ecd