bindings: soc: qcom: add Glink PMIC debug controller binding documentation
Add binding documentation for the Qualcomm Technologies, Inc.
Glink PMIC debug controller device.  This device provides an
interface to read and write PMIC registers over PMIC Glink using
a remote subsytem (e.g. DSP).  This allows for debugging PMIC
peripherals that would typically only be accessible to the
charger and fuel gauging firmware running on the remote
subsystem.  The PMICs may be physically connected using either
SPMI or I2C.

This is a snapshot of qcom,pmic-glink-debug.txt taken as of qcom-6.1
commit 5c116d156aae ("bindings: pmic-glink-debug: Add SPMI bridge bus
support") which was then converted to yaml format.

Change-Id: Ie4b3e188feea28b9c80413b0397eb1658e03835f
Signed-off-by: David Collins <quic_collinsd@quicinc.com>