ARM: dts: qcom: Add uses_level_shifter for sun
When this "qcom,uses_level_shifter" DT entry is present
but "qcom,external-fb-clk" is not present, the SD card's
HS50 mode is run at 37.5MHz clock. This is to work around
a latency issue caused by the addition of the Level Shifter
logic. In order to use the 50MHz clock in HS50 mode, there
are two ways:
- Keep "qcom,uses_level_shifter" and "qcom,external-fb-clk"
  and make sure the Feedback Clock (FB CLK) connection between
  the Level Shifter and the MSM must be good.

- Remove "qcom,uses_level_shifter" from DT.

Please note that the "qcom,external-fb-clk" depends on
"qcom,uses_level_shifter", only set "qcom,external-fb-clk"
would not take effect.

Change-Id: Ib3a4cf5c9c20e2094556c3196d91b7e9723e0a1e
Signed-off-by: Ziqi Chen <quic_ziqichen@quicinc.com>