ARM: dts: msm: Configure SD card_detect gpio to bias-pull-up in off state
When SD card is not present, card detect gpio status should be pulled high.

With bias-disable configuration, line is not pulled up as a result IRQ
storm for card insert is seen on the SD card detect line.

Configure SD card_detect gpio to bias-pull-up in off state on blair
platform.

Change-Id: I6b0712dbe391a8befd68799865ec5c2915ecdcaa