ARM: dts: msm: Reduce spm level to 3
We observed rpmb read failures with anorak with high spm level,
same is applicable to upgraded blair too. As mitigation need a
huge design change, its not feasilbe here to make that change.
Hence reduce spm level to 3.
Moreover this is inline with upstream way and PT targets.

Remove the entry from dt as default value taken is 3.

Change-Id: I59df49a3e15440663f5567f20badb518a93bfd59