ARM: dts: msm: Reduce spm level to 3
We observed rpmb read failures with anorak with high spm level,
same is applicable to upgraded blair too. As mitigation need a
huge design change, its not feasilbe here to make that change.
Hence reduce spm level to 3. Moreover this is in line with upstream
way and PT targets.

Remove the entries from dt as default value taken is 3.

Change-Id: Ie789d69dabfd800e0d02d33781b686c3621664bb