ARM: dts: qcom: add SPMI bus controller for Sun
Add spmi-pmic-arb devices for the primary and secondary SPMI buses
found on Sun.  The primary bus operates at 19.2 MHz and is used for
most of the PMICs.

The secondary bus operates at 4.8 MHz and is used exclusively for
charging PMICs.  Note that the secondary bus is not connected to
the SoC on the board due to voltage level differences.  Therefore,
keep the secondary bus device disabled.

Change-Id: I6b2bb6b54e285fd9c333971b08134c3768087869
Signed-off-by: David Collins <quic_collinsd@quicinc.com>