ARM: dts: msm: Configure SD card_detect gpio to bias-pull-up in off state
When SD card is not present, card detect gpio status should be pulled high.

With bias-disable configuration, line is not pulled up as a result IRQ
storm for card insert is seen on the SD card detect line.

Configure SD card_detect gpio to bias-pull-up in off state on holi
platform.

Change-Id: I8ad78f3c60a81c2d115e18d3fd17aa7831fe92b1