ARM: dts: qcom: Add PMIC ID for holi PM6350 variants
Holi supports 2 HW board design options:-
1. PM6125 as primary PMIC
2. PM6350 as primary PMIC

The DT settings for the primary PMIC are picked dynamically
based on the PMIC-ID defined in the DT overlay.

However, currently only Holi-PM6125 DT overlay has the
PMIC-ID defined. This is causing compilation issues
on the Bazel builds. Fix this by defining the PMIC-ID
property in the Holi-PM6350 DT overlay as well.

Change-Id: Iccdd3a8c139a0e26bdb5432ff8087434ae69224d