ARM: dts: msm: Add probe dependency to PMIC PON Log driver
Currently, the device doesn’t warm-reset when panic() is triggered by
the pmic-pon-log driver during probe() when a fault in PMIC HW is
detected and "qcom,pmic-fault-panic" property is set. This is because
gh_virt_wdt driver is not probed yet.

Fix this by ensuring that the virtual watchdog driver probes _before_
the PMIC PON log driver does.

Change-Id: Ied0a6b8221e3be62796f2df64e09c16bc232a499