ARM: dts: qcom: switch to RPMh control of VRM regulators on Sun
Replace the stub-regulator devices for VRM managed PMIC regulators
with rpmh-regulator devices.  This ensures that consumers are able
to modify the physical state of these PMIC regulators.

Update all regulator voltage limits to match the latest hardware
guidance.

Remove references to regulator L3B since RPMh will not be providing
support for it.

Change-Id: Ieee28b3860d013837ca273c32606bf3e0b0d2a6b
Signed-off-by: David Collins <quic_collinsd@quicinc.com>