dt-bindings: pci: Add device_type, cesta-l1sub-timeout-ext-int property
Add the device_type property in the pci device tree as a mandatory
property so that the ranges property will be read correctly by
the of framework (of.c) file when pci framework is parsing the
pci device tree nodes.

Also add the qcom,cesta-l1sub-timeout-ext-int property as an
optional property for the platforms that have cesta clkreq routing
bit disabled and legacy l1ss timeout enabled by default.
The presence of this property on a cesta enabled platform will
enable routing of the L1ss timeout interrupt to cesta HW block
instead of routing it to APPS subsystem.

Change-Id: I96db05d6cd38f507c7ddba968ee65808b8263076
Signed-off-by: Prudhvi Yarlagadda <quic_pyarlaga@quicinc.com>