ARM: dts: qcom: add gh-watchdog dependency to pmic-pon-log device for Sun
Add a dependency between the gh-watchdog and pmic-pon-log devices
for Sun boards.  This ensures that during system boot-up, the driver
responsible for the gh-watchdog device initializes before the one
responsible for the pmic-pon-log device.

This behavior is critical to ensuring that a PMIC warm-reset correctly
occurs in the situation that the qcom,pmic-fault-panic property is
specified for the pmic-pon-log device and the previous system
power-off was caused by a PMIC fault.

Change-Id: I8c40b19ce150ca5bb3cb9514c7d293481a5009af
Signed-off-by: David Collins <quic_collinsd@quicinc.com>