ARM: dts: msm: Add wcd_usbss reference to USB on sun
The USB D+/D- signal lines are first routed through the WCD939x USB
subsystem before connecting to the USB controller on MTP platform for
sun. Add a phandle to the former to the USB device node. This will allow
the USB driver to control the D+/D- switches when USB 2.0 functionality
is needed.

Change-Id: Iae9cc1fc46826c1f355e230029d6b5a41eae057f
Signed-off-by: Ronak Vijay Raheja <quic_rraheja@quicinc.com>