ARM: dts: msm: Update pmic support for volcano
Fix gpio config for pm8010 and slave id for pmiv0104.

Remove NON-PORED ldo nodes and ldos which
are not exposed to HLOS from volcano regulators, and
update the regulator type property for one ldo. Update
init mode for VDD_MODEM rail to retention as it doesn't
support turbo mode.

Remove pmr735a support as it is not present on target currently.

Change-Id: I93782c25a30d0b4672bdd79720eeae86bf757316