ARM: dts: msm: SA6155: Set PCIe boot-option to 0x1
PCIe root complex device (0000:00:00:00) PM OPs is causing 1sec
delay during LPM Resume. In order to avoid the PM resume/suspend
for device, driver can mark the device as syscore,so the PM driver
will skip the suspend/resume for particular device.
With device like WLAN which are externally powered resume needs to be
delayed till the device gets power.
So,set the boot-option as 0x1 to fix this issue.

Change-Id: Ib5fbe9d2978f61fcde6630cce1f74766a7c82ad1